Reduce the scope/deps of some Validator functions
- `charset_validator()`         -> `config/charset.c`
- `compress_method_validator()` -> `hcache/config.c`
- `hcache_validator()`          -> `hcache/config.c`
- `multipart_validator()`       -> `mutt_config.c`
- `pager_validator()`           -> `mutt_config.c`
- `reply_validator()`           -> `mutt_config.c`
- Add flags to `charset_validator()`

Closes: #2625
