+++ init.c
@@ -1543,165 +1543,3 @@ int mutt_var_value_complete(char *buf, size_t buflen, int pos)
   }
   return 0;
 }
-
-/**
- * charset_validator - Validate the "charset" config variable - Implements ConfigDef::validator()
- */
-int charset_validator(const struct ConfigSet *cs, const struct ConfigDef *cdef,
-                      intptr_t value, struct Buffer *err)
-{
-  if (value == 0)
-    return CSR_SUCCESS;
-
-  const char *str = (const char *) value;
-
-  if ((strcmp(cdef->name, "charset") == 0) && strchr(str, ':'))
-  {
-    mutt_buffer_printf(
-        err, _("'charset' must contain exactly one character set name"));
-    return CSR_ERR_INVALID;
-  }
-
-  int rc = CSR_SUCCESS;
-  bool strict = (strcmp(cdef->name, "send_charset") == 0);
-  char *q = NULL;
-  char *s = mutt_str_dup(str);
-
-  for (char *p = strtok_r(s, ":", &q); p; p = strtok_r(NULL, ":", &q))
-  {
-    if (*p == '\0')
-      continue;
-    if (!mutt_ch_check_charset(p, strict))
-    {
-      rc = CSR_ERR_INVALID;
-      mutt_buffer_printf(err, _("Invalid value for option %s: %s"), cdef->name, p);
-      break;
-    }
-  }
-
-  FREE(&s);
-  return rc;
-}
-
-#ifdef USE_HCACHE
-/**
- * hcache_validator - Validate the "header_cache_backend" config variable - Implements ConfigDef::validator()
- */
-int hcache_validator(const struct ConfigSet *cs, const struct ConfigDef *cdef,
-                     intptr_t value, struct Buffer *err)
-{
-  if (value == 0)
-    return CSR_SUCCESS;
-
-  const char *str = (const char *) value;
-
-  if (store_is_valid_backend(str))
-    return CSR_SUCCESS;
-
-  mutt_buffer_printf(err, _("Invalid value for option %s: %s"), cdef->name, str);
-  return CSR_ERR_INVALID;
-}
-
-#ifdef USE_HCACHE_COMPRESSION
-/**
- * compress_method_validator - Validate the "header_cache_compress_method" config variable - Implements ConfigDef::validator()
- */
-int compress_method_validator(const struct ConfigSet *cs, const struct ConfigDef *cdef,
-                              intptr_t value, struct Buffer *err)
-{
-  if (value == 0)
-    return CSR_SUCCESS;
-
-  const char *str = (const char *) value;
-
-  if (compress_get_ops(str))
-    return CSR_SUCCESS;
-
-  mutt_buffer_printf(err, _("Invalid value for option %s: %s"), cdef->name, str);
-  return CSR_ERR_INVALID;
-}
-/**
- * compress_level_validator - Validate the "header_cache_compress_level" config variable - Implements ConfigDef::validator()
- */
-int compress_level_validator(const struct ConfigSet *cs, const struct ConfigDef *cdef,
-                             intptr_t value, struct Buffer *err)
-{
-  if (!C_HeaderCacheCompressMethod)
-  {
-    mutt_buffer_printf(err, _("Set option %s before setting %s"),
-                       "header_cache_compress_method", cdef->name);
-    return CSR_ERR_INVALID;
-  }
-
-  const struct ComprOps *cops = compress_get_ops(C_HeaderCacheCompressMethod);
-  if (!cops)
-  {
-    mutt_buffer_printf(err, _("Invalid value for option %s: %s"),
-                       "header_cache_compress_method", C_HeaderCacheCompressMethod);
-    return CSR_ERR_INVALID;
-  }
-
-  if ((value < cops->min_level) || (value > cops->max_level))
-  {
-    // L10N: This applies to the "$header_cache_compress_level" config variable.
-    //       It shows the minimum and maximum values, e.g. 'between 1 and 22'
-    mutt_buffer_printf(err, _("Option %s must be between %d and %d inclusive"),
-                       cdef->name, cops->min_level, cops->max_level);
-    return CSR_ERR_INVALID;
-  }
-
-  return CSR_SUCCESS;
-}
-#endif /* USE_HCACHE_COMPRESSION */
-#endif /* USE_HCACHE */
-
-/**
- * pager_validator - Check for config variables that can't be set from the pager - Implements ConfigDef::validator()
- */
-int pager_validator(const struct ConfigSet *cs, const struct ConfigDef *cdef,
-                    intptr_t value, struct Buffer *err)
-{
-  if (CurrentMenu == MENU_PAGER)
-  {
-    mutt_buffer_printf(err, _("Option %s may not be set or reset from the pager"),
-                       cdef->name);
-    return CSR_ERR_INVALID;
-  }
-
-  return CSR_SUCCESS;
-}
-
-/**
- * multipart_validator - Validate the "show_multipart_alternative" config variable - Implements ConfigDef::validator()
- */
-int multipart_validator(const struct ConfigSet *cs, const struct ConfigDef *cdef,
-                        intptr_t value, struct Buffer *err)
-{
-  if (value == 0)
-    return CSR_SUCCESS;
-
-  const char *str = (const char *) value;
-
-  if (mutt_str_equal(str, "inline") || mutt_str_equal(str, "info"))
-    return CSR_SUCCESS;
-
-  mutt_buffer_printf(err, _("Invalid value for option %s: %s"), cdef->name, str);
-  return CSR_ERR_INVALID;
-}
-
-/**
- * reply_validator - Validate the "reply_regex" config variable - Implements ConfigDef::validator()
- */
-int reply_validator(const struct ConfigSet *cs, const struct ConfigDef *cdef,
-                    intptr_t value, struct Buffer *err)
-{
-  if (pager_validator(cs, cdef, value, err) != CSR_SUCCESS)
-    return CSR_ERR_INVALID;
-
-  if (!OptAttachMsg)
-    return CSR_SUCCESS;
-
-  mutt_buffer_printf(err, _("Option %s may not be set when in attach-message mode"),
-                     cdef->name);
-  return CSR_ERR_INVALID;
-}
